+++ frontend/components/commons/forms/questions-form/QuestionsForm.component.vue
@@ -24,49 +24,54 @@
       </div>
       <div class="form-group" v-for="input in inputs" :key="input.id">
         <TextAreaComponent
-          v-if="input.component_type === COMPONENT_TYPE.FREE_TEXT"
+          v-if="input.isTextType"
           :title="input.question"
           :placeholder="input.placeholder"
           v-model="input.options[0].value"
           :useMarkdown="input.settings.use_markdown"
-          :isRequired="input.is_required"
+          :hasSuggestion="!isRecordSubmitted && input.matchSuggestion"
+          :isRequired="input.isRequired"
           :description="input.description"
           @on-error="onError"
         />
 
         <SingleLabelComponent
-          v-if="input.component_type === COMPONENT_TYPE.SINGLE_LABEL"
+          v-if="input.isSingleLabelType"
           :questionId="input.id"
           :title="input.question"
           v-model="input.options"
-          :isRequired="input.is_required"
+          :hasSuggestion="!isRecordSubmitted && input.matchSuggestion"
+          :isRequired="input.isRequired"
           :description="input.description"
           :visibleOptions="input.settings.visible_options"
         />
 
         <MultiLabelComponent
-          v-if="input.component_type === COMPONENT_TYPE.MULTI_LABEL"
+          v-if="input.isMultiLabelType"
           :questionId="input.id"
           :title="input.question"
           v-model="input.options"
-          :isRequired="input.is_required"
+          :hasSuggestion="!isRecordSubmitted && input.matchSuggestion"
+          :isRequired="input.isRequired"
           :description="input.description"
           :visibleOptions="input.settings.visible_options"
         />
 
         <RatingComponent
-          v-if="input.component_type === COMPONENT_TYPE.RATING"
+          v-if="input.isRatingType"
           :title="input.question"
           v-model="input.options"
-          :isRequired="input.is_required"
+          :hasSuggestion="!isRecordSubmitted && input.matchSuggestion"
+          :isRequired="input.isRequired"
           :description="input.description"
           @on-error="onError"
         />
 
         <RankingComponent
-          v-if="input.component_type === COMPONENT_TYPE.RANKING"
+          v-if="input.isRankingType"
           :title="input.question"
-          :isRequired="input.is_required"
+          :hasSuggestion="!isRecordSubmitted && input.matchSuggestion"
+          :isRequired="input.isRequired"
           :description="input.description"
           v-model="input.options"
         />
@@ -128,6 +133,7 @@ import {
   upsertRecordResponses,
   deleteRecordResponsesByUserIdAndResponseId,
 } from "@/models/feedback-task-model/record-response/recordResponse.queries";
+import { useQuestionFormViewModel } from "./useQuestionsFormViewModel";
 
 export default {
   name: "QuestionsFormComponent",
@@ -144,17 +150,17 @@ export default {
       type: String,
       required: true,
     },
-    initialInputs: {
-      type: Array,
-      required: true,
-    },
   },
   data() {
     return {
       inputs: [],
+      initialInputs: [],
       renderForm: 0,
       isError: false,
     };
+  },
+  setup() {
+    return useQuestionFormViewModel();
   },
   computed: {
     userId() {
@@ -176,11 +182,11 @@ export default {
       const requiredQuestionsAreCompletedCorrectly = this.inputs
         .filter((input) => input.is_required)
         .every((input) => {
-          if (input.component_type === COMPONENT_TYPE.FREE_TEXT) {
-            return input.options[0]?.value.trim() != "";
+          if (input.isTextType) {
+            return input.options[0]?.value.trim() !== "";
           }
 
-          if (input.component_type === COMPONENT_TYPE.RANKING) {
+          if (input.isRankingType) {
             return input.options.every((option) => option.rank);
           }
 
@@ -190,7 +196,7 @@ export default {
       const optionalQuestionsCompletedAreCorrectlyEntered = this.inputs
         .filter((input) => !input.is_required)
         .every((input) => {
-          if (input.component_type === COMPONENT_TYPE.RANKING) {
+          if (input.isRankingType) {
             return (
               !input.options.some((option) => option.rank) ||
               input.options.every((option) => option.rank)
@@ -220,18 +226,18 @@ export default {
     currentInputsWithNoResponses() {
       return this.inputs.filter((input) => {
         if (
-          input.component_type === COMPONENT_TYPE.RATING ||
-          input.component_type === COMPONENT_TYPE.SINGLE_LABEL ||
-          input.component_type === COMPONENT_TYPE.MULTI_LABEL
+          input.isRatingType ||
+          input.isSingleLabelType ||
+          input.isMultiLabelType
         ) {
           return input.options.every((option) => !option.is_selected);
         }
 
-        if (input.component_type === COMPONENT_TYPE.RANKING) {
+        if (input.isRankingType) {
           return input.options.every((option) => !option.rank);
         }
 
-        if (input.component_type === COMPONENT_TYPE.FREE_TEXT) {
+        if (input.isTextType) {
           return !input.options[0]?.value.trim();
         }
       });
@@ -289,13 +295,21 @@ export default {
             responseValues
           );
         }
-        const { data: updatedResponses } = responseData;
 
-        if (updatedResponses) {
-          this.updateResponsesInOrm({
-            record_id: this.recordId,
-            ...updatedResponses,
-          });
+        const { data: updatedResponse } = responseData;
+        const answer = {
+          record_id: this.recordId,
+          ...updatedResponse,
+        };
+
+        if (this.responseId) {
+          this.updateResponse(answer);
+        } else {
+          this.addResponse(answer);
+        }
+
+        if (updatedResponse) {
+          this.updateResponsesInOrm(answer);
         }
       } catch (error) {
         console.log(error);
@@ -321,7 +335,6 @@ export default {
 
         this.$emit("on-discard-responses");
 
-        // TODO - reset only when we know that we fetch and computed all the necessary records data
         this.onReset();
       } catch (error) {
         console.log(error);
@@ -349,7 +362,6 @@ export default {
 
         this.$emit("on-submit-responses");
 
-        // TODO - reset only when we know that we fetch and computed all the necessary records data
         this.onReset();
       } catch (error) {
         console.log(error);
@@ -372,13 +384,28 @@ export default {
           RECORD_STATUS.PENDING
         );
 
+        this.clearRecord(this.recordId, RECORD_STATUS.PENDING);
+
         this.$emit("on-clear-responses");
-        this.onReset();
+
+        if (this.responseId) {
+          this.$nextTick(() => {
+            this.onReset();
+          });
+        } else {
+          this.$nextTick(() => {
+            this.initialInputs = this.feedback.getAnswerWithNoSuggestions();
+            this.inputs = cloneDeep(this.initialInputs);
+            this.isError = false;
+            this.renderForm++;
+          });
+        }
       } catch (err) {
         console.log(err);
       }
     },
     onReset() {
+      this.initialInputs = this.feedback.getAnswer(this.recordId, this.userId);
       this.inputs = cloneDeep(this.initialInputs);
       this.isError = false;
       this.renderForm++;
@@ -502,7 +529,7 @@ export default {
       return formattedRecordResponsesForOrm;
     },
     factoryInputsToResponseValues() {
-      let responseByQuestionName = {};
+      const responseByQuestionName = {};
 
       this.inputs.forEach((input) => {
         switch (input.component_type) {
@@ -521,7 +548,6 @@ export default {
           case COMPONENT_TYPE.RATING: {
             const selectedOption =
               input.options?.find((option) => option.is_selected) ?? false;
-
             if (selectedOption) {
               responseByQuestionName[input.name] = {
                 value: selectedOption.value,
@@ -531,7 +557,6 @@ export default {
           }
           case COMPONENT_TYPE.FREE_TEXT: {
             const text = input.options[0]?.value.trim();
-
             if (text) {
               responseByQuestionName[input.name] = {
                 value: text,