+++ frontend/specs/components/filters/SortList.spec.js
@@ -36,7 +36,7 @@ function mountSortList() {
 }
 
 describe("SortList", () => {
-  let spy = jest.spyOn(console, "error");
+  const spy = jest.spyOn(console, "error");
   afterEach(() => spy.mockReset());
 
   test("renders properly", () => {