+++ src/argilla/client/users.py
@@ -26,12 +26,14 @@ from argilla.client.sdk.commons.errors import (
 from argilla.client.sdk.users import api as users_api
 from argilla.client.sdk.users.models import UserCreateModel, UserModel, UserRole
 from argilla.client.sdk.v1.users import api as users_api_v1
-from argilla.client.sdk.v1.workspaces.models import WorkspaceModel
+from argilla.client.sdk.v1.workspaces import api as workspaces_api_v1
+from argilla.client.utils import allowed_for_roles
 
 if TYPE_CHECKING:
     import httpx
 
     from argilla.client.sdk.client import AuthenticatedClient
+    from argilla.client.sdk.v1.workspaces.models import WorkspaceModel
 
 
 class User:
@@ -59,7 +61,7 @@ class User:
         >>> from argilla import rg
         >>> user = rg.User.from_name("my-user") # or `User.from_id("...")`
         >>> print(user)
-        User(id='...', username='my-user', first_name='Luke', last_name="Skywalker', full_name='Luke Skywalker', role='annotator', workspaces=[WorkspaceModel(...), ...], api_key='...', inserted_at=datetime.datetime(2021, 8, 31, 10, 0, 0), updated_at=datetime.datetime(2021, 8, 31, 10, 0, 0))
+        User(id='...', username='my-user', role='annotator', first_name='Luke', last_name="Skywalker', full_name='Luke Skywalker', role='annotator', api_key='...', inserted_at=datetime.datetime(2021, 8, 31, 10, 0, 0), updated_at=datetime.datetime(2021, 8, 31, 10, 0, 0))
     """
 
     __client: "httpx.Client"
@@ -107,20 +109,23 @@ class User:
         raise Exception(error_msg)
 
     @property
-    def workspaces(self) -> Optional[List[WorkspaceModel]]:
+    def workspaces(self) -> Optional[List["WorkspaceModel"]]:
         """Returns the workspace names the current user is linked to.
 
         Returns:
             A list of `WorkspaceModel` the current user is linked to.
         """
-        return users_api_v1.list_user_workspaces(self.__client, self.id).parsed
+        connected_user = users_api.whoami_httpx(self.__client).parsed
+        if connected_user.role == UserRole.owner:
+            return users_api_v1.list_user_workspaces(self.__client, self.id).parsed
+        return workspaces_api_v1.list_workspaces_me(self.__client).parsed
 
     def __repr__(self) -> str:
         return (
             f"User(id={self.id}, username={self.username}, role={self.role},"
-            f" workspaces={self.workspaces}, api_key={self.api_key},"
-            f" first_name={self.first_name}, last_name={self.last_name}, role={self.role},"
-            f" inserted_at={self.inserted_at}, updated_at={self.updated_at})"
+            f" api_key={self.api_key}, first_name={self.first_name},"
+            f" last_name={self.last_name}, inserted_at={self.inserted_at},"
+            f" updated_at={self.updated_at})"
         )
 
     @staticmethod
@@ -134,6 +139,7 @@ class User:
         except Exception as e:
             raise RuntimeError(f"The `rg.active_client()` is not available or not respoding.") from e
 
+    @allowed_for_roles(roles=[UserRole.owner])
     def delete(self) -> None:
         """Deletes the user from Argilla.
 
@@ -165,6 +171,7 @@ class User:
         return instance
 
     @classmethod
+    @allowed_for_roles(roles=[UserRole.owner])
     def create(
         cls,
         username: str,
@@ -225,6 +232,7 @@ class User:
             raise RuntimeError(f"Error while creating user with username=`{username}` in Argilla.") from e
 
     @classmethod
+    @allowed_for_roles(roles=[UserRole.owner])
     def from_id(cls, id: UUID) -> "User":
         """Gets an existing user from Argilla by its ID.
 
@@ -257,6 +265,7 @@ class User:
             raise RuntimeError(f"Error while retrieving user with id=`{id}` from Argilla.") from e
 
     @classmethod
+    @allowed_for_roles(roles=[UserRole.owner])
     def from_name(cls, name: str) -> "User":
         """Gets an existing user from Argilla by its name.
 
@@ -316,6 +325,7 @@ class User:
             raise RuntimeError("Error while retrieving the current user from Argilla.") from e
 
     @classmethod
+    @allowed_for_roles(roles=[UserRole.owner])
     def list(cls) -> Iterator["User"]:
         """Lists all the users in Argilla.