+++ tests/client/sdk/v1/users/test_api.py
@@ -13,7 +13,6 @@
 #  limitations under the License.
 
 from typing import TYPE_CHECKING
-from uuid import uuid4
 
 import pytest
 from argilla.client.api import ArgillaSingleton